feat: add training module
